refactor(my-key-results): improves file structure
