+++ src/components/Cycle/Filter/QuarterSelector/index.ts
@@ -1,0 +1,3 @@
+import CycleFilterQuarterSelector from './quarter-selector'
+
+export default CycleFilterQuarterSelector
