+++ src/components/Cycle/Filter/YearSelector/index.ts
@@ -1,0 +1,3 @@
+import CycleFilterYearSelector from './year-selector'
+
+export default CycleFilterYearSelector
