+++ src/components/Cycle/Filter/YearSelector/messages.ts
@@ -5,10 +5,6 @@ type CycleFilterMessage =
   | 'yearLeftIconDesc'
   | 'yearRightIconDescClosed'
   | 'yearRightIconDescOpened'
-  | 'quarterEmptyStateFirstButton'
-  | 'quarterEmptyStateSecondButton'
-  | 'quarterEmptyStateThirdButton'
-  | 'quarterEmptyStateFourthButton'
 
 export default defineMessages<CycleFilterMessage>({
   yearSelectorEmptyState: {
@@ -36,33 +32,5 @@ export default defineMessages<CycleFilterMessage>({
     id: '7d+540',
     description:
       'This message is displayed to the user screen reader when the year select menu is opened',
-  },
-
-  quarterEmptyStateFirstButton: {
-    defaultMessage: 'Q1',
-    id: 'fI4P80',
-    description:
-      'This button appears as an empty state while the user is selecting the cycle filter',
-  },
-
-  quarterEmptyStateSecondButton: {
-    defaultMessage: 'Q2',
-    id: 'aKyktL',
-    description:
-      'This button appears as an empty state while the user is selecting the cycle filter',
-  },
-
-  quarterEmptyStateThirdButton: {
-    defaultMessage: 'Q3',
-    id: 'znpwST',
-    description:
-      'This button appears as an empty state while the user is selecting the cycle filter',
-  },
-
-  quarterEmptyStateFourthButton: {
-    defaultMessage: 'Q4',
-    id: 'u4rqek',
-    description:
-      'This button appears as an empty state while the user is selecting the cycle filter',
   },
 })
