+++ src/components/Cycle/Filter/YearSelector/queries.gql
@@ -1,0 +1,6 @@
+query GET_NOT_ACTIVE_YEARLY_CYCLES {
+  cycles(active: false, cadence: YEARLY) {
+    id
+    title
+  }
+}
